test: give collection-update events a dedicated lane suite
