+++ packages/publisher/test/chain-event-lane-publish.test.ts
@@ -5,8 +5,6 @@ import { OxigraphStore } from '@origintrail-official/dkg-storage';
 import { describe, expect, it } from 'vitest';
 import type { ChainEvent } from '@origintrail-official/dkg-chain';
 import { ChainEventPoller } from '../src/chain-event-poller.js';
-import type { ChainEventPollerLane } from '../src/chain-event-poller.js';
-import type { LaneCursorPersistence } from '../src/chain-event-poller.js';
 import { PublishHandler } from '../src/publish-handler.js';
 import { makeChain, makeHandler, markPending, journalEntry } from './helpers/chain-event-lane-fixture.js';
 
@@ -237,41 +235,5 @@ describe('ChainEventPoller publish', () => {
     expect(filters[1].fromBlock).toBe(1_991_001);
     expect(filters[1].toBlock).toBe(2_000_000);
     expect(confirmed).toEqual([0x11, 0x22]);
-  });
-
-  it('dispatches collection update events and advances the collectionUpdates cursor', async () => {
-    const event: ChainEvent = {
-      type: 'KnowledgeAssetUpdated',
-      blockNumber: 50,
-      data: {
-        merkleRoot: '0x' + '44'.repeat(32),
-        batchId: '42',
-      },
-    };
-    const { adapter, filters } = makeChain(100, [event]);
-    const saveCalls: Array<{ lane: ChainEventPollerLane; block: number }> = [];
-    const cursor: LaneCursorPersistence = {
-      async loadLane() { return undefined; },
-      async saveLane(lane, block) { saveCalls.push({ lane, block }); },
-    };
-    const seen: Array<{ merkleRoot: Uint8Array; batchId: bigint; blockNumber: number }> = [];
-    const poller = new ChainEventPoller({
-      chain: adapter,
-      publishHandler: makeHandler(),
-      intervalMs: 60_000,
-      cursorPersistence: cursor,
-      onCollectionUpdated: async (info) => { seen.push(info); },
-    });
-
-    await poller.start();
-    await new Promise((r) => setTimeout(r, 50));
-    await poller.stop();
-
-    expect(filters[0].eventTypes).toEqual(['KnowledgeAssetUpdated']);
-    expect(seen).toHaveLength(1);
-    expect(Buffer.from(seen[0].merkleRoot).toString('hex')).toBe('44'.repeat(32));
-    expect(seen[0].batchId).toBe(42n);
-    expect(seen[0].blockNumber).toBe(50);
-    expect(saveCalls).toEqual([{ lane: 'collectionUpdates', block: 100 }]);
   });
 });
